nav bar fixes with logout button
