+++ src/components/Navigation/Navigation.css
@@ -33,8 +33,8 @@
 .navigation__links {
   display: flex;
   gap: 24px;
-  flex: 1 1 auto;
-  justify-content: flex-start;
+  align-items: center;
+  margin-right: 16px;
 }
 
 .navigation__link {
@@ -49,7 +49,7 @@
 }
 
 .navigation__link--saved {
-  margin-right: 16px;
+  margin-right: 0;
 }
 
 .navigation__link:hover {
@@ -91,8 +91,18 @@
 
 .navigation__button_type_logout {
   background-color: transparent;
-  border: 1px solid #d1d2d6;
-  color: #1a1b22;
+  border: 1px solid #fff;
+  color: #fff;
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+
+.navigation__logout-icon {
+  width: 18px;
+  height: 18px;
+  margin-left: 2px;
+  filter: brightness(0) invert(1);
 }
 
 .navigation__button:hover {
@@ -111,15 +121,17 @@
 
   .navigation__links {
     gap: 16px;
-  }
-
-  .navigation__link--saved {
     margin-right: 12px;
   }
 
   .navigation__button {
     padding: 8px 16px;
     font-size: 16px;
+  }
+
+  .navigation__logout-icon {
+    width: 16px;
+    height: 16px;
   }
 }
 
@@ -135,9 +147,6 @@
 
   .navigation__links {
     gap: 12px;
-  }
-
-  .navigation__link--saved {
     margin-right: 10px;
   }
 
@@ -158,15 +167,12 @@
 
   .navigation__links {
     gap: 10px;
+    margin-right: 8px;
   }
 
   .navigation__link {
     font-size: 14px;
     padding: 8px 4px;
-  }
-
-  .navigation__link--saved {
-    margin-right: 8px;
   }
 
   .navigation__button {
@@ -176,6 +182,11 @@
 
   .navigation__buttons {
     gap: 8px;
+  }
+
+  .navigation__logout-icon {
+    width: 14px;
+    height: 14px;
   }
 }
 
@@ -191,16 +202,12 @@
 
   .navigation__links {
     gap: 8px;
+    margin-right: 6px;
   }
 
   .navigation__link {
     font-size: 13px;
     padding: 6px 3px;
-  }
-
-  .navigation__link--saved {
-    margin-right: 6px;
-    font-size: 12px;
   }
 
   .navigation__buttons {
@@ -212,16 +219,11 @@
     font-size: 13px;
   }
 
-  /* Stack navigation items vertically on very small screens */
+  /* Fix for very small screens */
   .navigation__right {
-    flex-direction: column;
-    align-items: flex-end;
-    gap: 4px;
-    margin-top: -20px; /* Adjust vertical position */
-  }
-
-  .navigation__link--saved {
-    margin-right: 0;
-    margin-bottom: 4px;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    margin-top: 0;
   }
 }
